requirements.txt
streamlit>=1.31.0
openai>=1.12.0
chromadb>=0.4.22
python-dotenv>=1.0.0
pyyaml>=6.0.0
tiktoken>=0.6.0
pandas>=2.0.0
